refactor : step3 폴더구조 변경
