+++ src/app/join/funnels/step3/Step3Component.client.tsx
@@ -1,6 +1,7 @@
 'use client';
 
-import FormSection from './components/FormSection.client';
+import CertificationForm from './components/CertificationForm.client';
+import EmailForm from './components/EmailForm.client';
 import NoticeSection from './components/NoticeSection.client';
 import TimerContext from './components/TimerContext.client';
 import JoinTitleTextMessage from '../../components/JoinTitleTextMessage.client';
@@ -17,7 +18,8 @@ export default function Step3Component() {
 
       <ModalContextProvider>
         <TimerContext>
-          <FormSection />
+          <EmailForm />
+          <CertificationForm />
         </TimerContext>
       </ModalContextProvider>
