--- src/app/join/funnels/step3/components/FormSection.client.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import CertificationForm from './CertificationForm.client';
-import EmailForm from './EmailForm.client';
-
-export default function InputSection() {
-  return (
-    <section>
-      <EmailForm />
-      <CertificationForm />
-    </section>
-  );
-}
